Add tags for people
The domain model around who's a civil servant, who's a contractor, who's
on loan from which department is quite complex and GDS specific. I'd
like to keep this application somewhat abstract, so it's more generally
useful, so I'm going with a kind of tag based approach.

The idea will be that you create a tag like `contractor` which you can
then apply to people on the team who are contractors, so you can see
them at a glance. In the current people plan spreadsheet, we sometimes
indicate a contractor with italic text.

I might also expand tags to other bits of the model, for example
representing Tech Leads as a `tech_lead` tag on a TeamAssignment.